% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{festival_data}
\alias{festival_data}
\title{Korean Festivals Data}
\format{
A data frame with 27 variables:
\describe{
  \item{\code{festival_kr}}{The name of the festival in Korean.}
  \item{\code{festival_en}}{The English translation of the festival's name.}
  \item{\code{place}}{The place where the festival is held, in Korean.}
  \item{\code{place_en}}{The English translation of the festival's location.}
  \item{\code{begin_date}}{The start date of the festival.}
  \item{\code{end_date}}{The end date of the festival.}
  \item{\code{contents}}{Description of the festival in Korean.}
  \item{\code{contents_en}}{The English translation of the festival's description.}
  \item{\code{organizer}}{The organizer of the festival in Korean.}
  \item{\code{organizer_en}}{The English translation of the organizer's name.}
  \item{\code{address}}{The address of the festival location in Korean.}
  \item{\code{address_en}}{The English translation of the address.}
  \item{\code{address_old_style}}{The old-style address in Korean.}
  \item{\code{address_old_en}}{The English translation of the old-style address.}
  \item{\code{latitude}}{Latitude coordinate of the festival location.}
  \item{\code{longitude}}{Longitude coordinate of the festival location.}
  \item{\code{updated}}{The date when the data was last updated.}
  \item{\code{data_organization_code}}{Code of the organization providing the data.}
  \item{\code{data_organization}}{The organization providing the data, in Korean.}
  \item{\code{data_organization_en}}{The English translation of the data organization's name.}
  \item{\code{other columns}}{Descriptions for any additional columns not listed here.}
}
}
\source{
\url{https://www.data.go.kr/data/15013104/standard.do}
}
\usage{
festival_data
}
\value{
Returns the raw data frame containing information on Korean festivals
}
\description{
\code{festival_data} returns the raw data frame containing
detailed information about festivals in Korea, including names, locations,
dates, and additional details in both Korean and English. English
translations are done through =googletranslate() on Google Sheets, and as
such may not be accurate.
}
\keyword{datasets}
