% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{international_students_all_individual_data}
\alias{international_students_all_individual_data}
\title{International Students Data (Individual)}
\format{
A data frame with 5 variables:
\describe{
  \item{\code{sex}}{The sex of the individual student.}
  \item{\code{university_kr}}{The name of the university the student is attending in Korean.}
  \item{\code{iso3c}}{The ISO 3-letter country code.}
  \item{\code{country}}{The English name of the student's country of origin.}
  \item{\code{type}}{The type of visa held by the student (e.g., Language, Degree).}
}
}
\source{
\url{https://www.data.go.kr/data/15100039/fileData.do}
}
\usage{
international_students_all_individual_data
}
\value{
Returns the raw data frame containing information on
international students in Korea.
}
\description{
\code{international_students_all_individual_data} A dataset
containing information about individual international students in Korea,
including their country of origin, university, visa type, and sex
of the students.
}
\keyword{datasets}
