% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{international_students_data}
\alias{international_students_data}
\title{International Students Data}
\format{
A data frame with 7 variables:
\describe{
  \item{\code{year}}{The year of the data record.}
  \item{\code{month}}{The month of the data record.}
  \item{\code{country_kr}}{The name of the student's country of origin in Korean.}
  \item{\code{visa}}{The type of visa held by the student (e.g., Language, Degree).}
  \item{\code{no_int_student}}{The number of international students from the respective country and visa type.}
  \item{\code{iso3c}}{The ISO 3-letter country code.}
  \item{\code{country}}{The English name of the student's country of origin.}
}
}
\source{
\url{https://www.data.go.kr/data/15100039/fileData.do}
}
\usage{
international_students_data
}
\value{
Returns the raw data frame containing information on Korean festivals
}
\description{
\code{international_students_data} A dataset containing information about
international students in Korea, including their country of origin,
visa type, and the number of students.
}
\keyword{datasets}
