% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kdata-package.R
\docType{_PACKAGE}
\name{kdata-package}
\alias{kdata-package}
\title{Korean Culture, Business, and Education Data}
\description{
Generate customized data frames containing various data regarding Korean
culture, business, and education.
}
\details{
The \code{kdata} package contains functions for generating various data
regarding Korean culture, business, and education.
}
\references{
\url{https://ayhan.phd/}
   \url{https://github.com/kjayhan/kdata}
}
\author{
Kadir Jun Ayhan \email{k@ayhan.phd}
}
\keyword{package}
