% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{status_of_foreign_students_data}
\alias{status_of_foreign_students_data}
\title{International Students Data (Schools)}
\format{
A data frame with 7 variables:
\describe{
 \item{\code{year}}{The year of the data record.}
 \item{\code{type_of_school}}{The type of school (e.g., university, cyber university).}
 \item{\code{classification}}{The classification of the school (e.g., private, public).}
 \item{\code{region}}{The region where the school is located.}
 \item{\code{school}}{The name of the school in Korean.}
 \item{\code{field}}{The field of study.}
 \item{\code{number_of_students}}{The number of international students at the school per field.}
}
}
\source{
\url{https://www.academyinfo.go.kr/uipnh/unt/unmcom/RdViewer.do?paramSvyYr=2023&paramMjrItem1=00&paramFormClftCd=30&paramItemId=33&paramSchlDivCd=02&paramDiv=A&paramItemDivCd=01&paramItemTimeCd=&paramItemDetailDivCd=&paramSchlKindCd=99&paramSchlEstabCd=99&paramZoneCd=99&paramSortItem=SCHL_NM&paramSortMethod=ASC&paramMjrCd=00&paramMjrItem2=00&paramMjrItem3=00&paramMjrItem4=00&paramMjrItem1Act=N&paramMjrItem4Act=N&paramMjrUsStp=0&paramSearchItem=SEARCH_SCHL_NM&tabIdx=&pageIdx=; https://www.data.go.kr/data/15050054/fileData.do}
}
\usage{
status_of_foreign_students_data
}
\value{
Returns the raw data frame containing information on
the current status of international students at Korean universities.
}
\description{
\code{status_of_foreign_students_data} A dataset containing
information about the number of international students at Korean
schools, including the school's name, region, the type of school
(university, cyber university, etc.), classification (private vs. public),
the field of study, and the number of students.
}
\keyword{datasets}
